Add_Mat_Vec and Mul_Mat_Vec are added to the ML.Mat/Vec.ecl
Some RECORD definitions are included in M_types which are for Matrix and Cell operations
(Here cell is a RECORD definition which includes and "id" and and child record set which is a "matrix"

cell function is implemneted to form a combined record out of a parent (which are ids)
and any number of children (which are matrix records and the id assigned to them)
this actually make a cell datastes that each record includes and id and the matrix assigned ot that id

BF_first3 is a very simple implementaion for the main loop in gradient descent algortihm
which is the feed forward pass and the update pass. I just made this implementation to test how I can
have a loop in ECL. my implementation is based on using ITERATE. it still need works to be done to become the gradient descent algorithm.