SoftMax_Sparse classifier is added to ML.Classify The implementation is based on the stanfrod's deep learning tutorial available at http://ufldl.stanford.edu/wiki/index.php/Softmax_Regression This implementation is based on using ML.Mat library. Test file is added to ML/Tests/Explanatory The implementation is tested on this toy dataset to compare the results with the results I get from my MATLAB implementation
Some functions are added to Mat.Has and Mat.Each

RandMat is added to ML.Mat that initializes a randome matrix

the fucntion ToGroundTruth is added to ML.Utils. This function converts the label
data to a specific matrix format for the label data that is used in the implementations.
The matrix is a "number of classes"*"number of samples" matrix.
each column is corresponded to one sample. All the lements are zero in one column
unless the element that it's row number is the same as the lable of that sample which
takes the value "1". in other word M(r, c) is 1 if y(c) = r and 0 otherwise