ready to test on lshtc large
softmax_lbfgs_lshtc_large8 includes the test for this dataset. The 8 is
there because I had an implementation ofr REAL4 and due to adding up a
large number of float numbers causing precision problem it didn't work.
As I changed all previous implementations to REAL4, I had to add an 8 to
the implementations of REAL8

I change the implementation of softmax in a way that parameter matrix is
partitioned row wise. tx is calculated by multiplying PBBlas format to
sparse format. This avoid converting data to PBblas format which
increases the memory usage due to adding a lot of zeros to the
partitions. Th eimplementation is in
deeplearning8.SoftMax_compatible_lbfgs_sparse_label_sparse

also the optimization that works for this function is
Optimization_new_new_2_2_nf_cost. I had to change cost function format
because the label data and train data are not represented as PBblas
anymore